Extract related content facets loading into separate service
