+++ bundle/API/Repository/TagsService.php
@@ -197,19 +197,6 @@ interface TagsService
     public function getRelatedContentCount(Tag $tag, array $additionalCriteria = []);
 
     /**
-     * Returns facets for given $facetBuilders,
-     * for content tagged with $tag.
-     *
-     * @param \Netgen\TagsBundle\API\Repository\Values\Tags\Tag $tag
-     * @param \eZ\Publish\API\Repository\Values\Content\Query\FacetBuilder[] $facetBuilders
-     *
-     * @throws \Netgen\TagsBundle\Exception\FacetingNotSupportedException
-     *
-     * @return \eZ\Publish\API\Repository\Values\Content\Search\Facet[]
-     */
-    public function getRelatedContentFacets(Tag $tag, array $facetBuilders = []);
-
-    /**
      * Creates the new tag.
      *
      * @param \Netgen\TagsBundle\API\Repository\Values\Tags\TagCreateStruct $tagCreateStruct
