+++ bundle/Resources/config/forms.yml
@@ -21,7 +21,7 @@ services:
         class: Netgen\TagsBundle\Form\Type\RelatedContentFilterType
         # public: false
         arguments:
-            - "@eztags.api.service.tags"
+            - "@eztags.repository.related_content_facets_loader"
             - "@ezpublish.api.service.content_type"
             - "@eztags.search.related_content.sort_service"
         tags:
